Refactor: move indexer code into a new indexer file.
